// backend/src/co-dong/dto/response/co-dong.response.dto.ts
export class CoDongResponseDto {
    id: number;

    hoTen: string;

    soDKSH: string;

    ngayCap: string[];

    diaChi: string;

    email: string;

    dienThoai: string;

    quocTich: string;

    slckngCong: number;

    ghiChu: string;

    type: number;

    cntc: number;

    txnum: string;

}